# test(sns): speed up integration tests that involve SNS archive even more