The new approach is to submit all the requests, then await all the responses, when spamming archive.

On my machine, the total runtime of upgrade_everything_test goes from `109.0s`-`126.8s` to `64.3s`-`65.9s` after this change.